+++ apps/app/src/components/decisions/proposalEditor/useRestoreProposalVersion.ts
@@ -92,7 +92,9 @@ export function useRestoreProposalVersion({
 
     const restoredData = buildRestoredProposalData(fragmentContents);
 
-    provider.revertToVersion(versionId, fragmentNames);
+    provider.revertToVersion(versionId, {
+      fields: fragmentNames,
+    });
 
     await updateProposalMutation.mutateAsync({
       proposalId,
