+++ packages/common/src/services/decision/updateDecisionInstance.ts
@@ -5,8 +5,6 @@ import {
   decisionProcessTransitions,
   processInstances,
   profiles,
-  taxonomies,
-  taxonomyTerms,
 } from '@op/db/schema';
 import { Events, event } from '@op/events';
 import type { User } from '@op/supabase/lib';
@@ -17,6 +15,7 @@ import { getProfileAccessUser } from '../access';
 import { assertProfileAdmin } from '../assert';
 import { generateUniqueProfileSlug } from '../profile/utils';
 import { createTransitionsForProcess } from './createTransitionsForProcess';
+import { ensureProposalTaxonomy } from './proposalTaxonomy';
 import { schemaValidator } from './schemaValidator';
 import type {
   DecisionInstanceData,
@@ -24,77 +23,6 @@ import type {
 } from './schemas/instanceData';
 import type { ProcessConfig } from './schemas/types';
 import { updateTransitionsForProcess } from './updateTransitionsForProcess';
-
-async function ensureProposalTaxonomy(categories: string[]): Promise<string[]> {
-  if (!categories || categories.length === 0) {
-    return [];
-  }
-
-  let proposalTaxonomy = await db._query.taxonomies.findFirst({
-    where: eq(taxonomies.name, 'proposal'),
-  });
-
-  if (!proposalTaxonomy) {
-    const [newTaxonomy] = await db
-      .insert(taxonomies)
-      .values({
-        name: 'proposal',
-        description:
-          'Categories for organizing proposals in decision-making processes',
-      })
-      .returning();
-
-    if (!newTaxonomy) {
-      throw new CommonError('Failed to create proposal taxonomy');
-    }
-    proposalTaxonomy = newTaxonomy;
-  }
-
-  const taxonomyTermIds: string[] = [];
-
-  for (const categoryName of categories) {
-    if (!categoryName.trim()) {
-      continue;
-    }
-
-    const categoryLabel = categoryName.trim();
-    const termUri = categoryLabel
-      .toLowerCase()
-      .replace(/\s+/g, '-')
-      .replace(/[^a-z0-9-]/g, '');
-
-    let existingTerm = await db._query.taxonomyTerms.findFirst({
-      where: eq(taxonomyTerms.termUri, termUri),
-    });
-
-    if (!existingTerm) {
-      const [newTerm] = await db
-        .insert(taxonomyTerms)
-        .values({
-          taxonomyId: proposalTaxonomy.id,
-          termUri,
-          label: categoryLabel,
-          definition: `Category for ${categoryLabel} proposals`,
-        })
-        .returning();
-
-      if (!newTerm) {
-        throw new CommonError(
-          `Failed to create taxonomy term for category: ${categoryLabel}`,
-        );
-      }
-      existingTerm = newTerm;
-    }
-
-    const taxonomyTerm = existingTerm;
-
-    if (taxonomyTerm) {
-      taxonomyTermIds.push(taxonomyTerm.id);
-    }
-  }
-
-  return taxonomyTermIds;
-}
 
 /**
  * Updates a decision process instance.
